`guardWith` parser
